+++ examples/nr-v2x-examples/nr-v2x-west-to-east-highway.cc
@@ -968,7 +968,7 @@ main(int argc, char* argv[])
             }
         }
         // all node ids, which are not in txSlUes container are Rx node ids
-        for (uint16_t i = 0; i < allSlUesContainer.GetN(); i++)
+        for (uint32_t i = 0; i < allSlUesContainer.GetN(); i++)
         {
             Ptr<Node> node = allSlUesContainer.Get(i);
             if (!txSlUes.Contains(node->GetId()))
@@ -1095,7 +1095,7 @@ main(int argc, char* argv[])
     double realAppStopTime = 0.0;
     double txAppDuration = 0.0;
 
-    for (uint16_t i = 0; i < txSlUes.GetN(); i++)
+    for (uint32_t i = 0; i < txSlUes.GetN(); i++)
     {
         clientApps.Add(sidelinkClient.Install(txSlUes.Get(i)));
         double jitter = startTimeSeconds->GetValue();
@@ -1121,7 +1121,7 @@ main(int argc, char* argv[])
     ApplicationContainer serverApps;
     PacketSinkHelper sidelinkSink("ns3::UdpSocketFactory", localAddress);
     sidelinkSink.SetAttribute("EnableSeqTsSizeHeader", BooleanValue(true));
-    for (uint16_t i = 0; i < rxSlUes.GetN(); i++)
+    for (uint32_t i = 0; i < rxSlUes.GetN(); i++)
     {
         serverApps.Add(sidelinkSink.Install(rxSlUes.Get(i)));
         serverApps.Start(Seconds(0.0));
@@ -1172,7 +1172,7 @@ main(int argc, char* argv[])
     if (!useIPv6)
     {
         // Set Tx traces
-        for (uint16_t ac = 0; ac < clientApps.GetN(); ac++)
+        for (uint32_t ac = 0; ac < clientApps.GetN(); ac++)
         {
             Ipv4Address localAddrs = clientApps.Get(ac)
                                          ->GetNode()
@@ -1189,7 +1189,7 @@ main(int argc, char* argv[])
         }
 
         // Set Rx traces
-        for (uint16_t ac = 0; ac < serverApps.GetN(); ac++)
+        for (uint32_t ac = 0; ac < serverApps.GetN(); ac++)
         {
             Ipv4Address localAddrs = serverApps.Get(ac)
                                          ->GetNode()
@@ -1208,7 +1208,7 @@ main(int argc, char* argv[])
     else
     {
         // Set Tx traces
-        for (uint16_t ac = 0; ac < clientApps.GetN(); ac++)
+        for (uint32_t ac = 0; ac < clientApps.GetN(); ac++)
         {
             clientApps.Get(ac)->GetNode()->GetObject<Ipv6L3Protocol>()->AddMulticastAddress(
                 groupAddress6);
@@ -1227,7 +1227,7 @@ main(int argc, char* argv[])
         }
 
         // Set Rx traces
-        for (uint16_t ac = 0; ac < serverApps.GetN(); ac++)
+        for (uint32_t ac = 0; ac < serverApps.GetN(); ac++)
         {
             serverApps.Get(ac)->GetNode()->GetObject<Ipv6L3Protocol>()->AddMulticastAddress(
                 groupAddress6);
